backend/requirements.txt
appdirs==1.4.3
attrs==18.2.0
backcall==0.1.0
black==18.9b0
bleach==3.1.0
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
Flask==1.0.2
Flask-Cors==3.0.7
google-api-core==1.7.0
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-speech==0.36.3
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.18.0
httplib2==0.12.1
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
PyMySQL==0.9.3
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
qtconsole==4.4.3
requests==2.21.0
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
toml==0.10.0
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1

